+++ src/components/Tooltip/index.ts
@@ -1,3 +1,3 @@
-// Tooltip from copy-ui @ 2025-09-25.
+// Tooltip from copy-ui @ 2025-09-30.
 
 export { Tooltip } from "./Tooltip";